+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "webpack": "webpack -p"
+    "webpack": "webpack -p",
+    "lint": "eslint --ignore-path .gitignore --fix ."
   },
   "repository": {
     "type": "git",
