+++ src/end.js
@@ -1,0 +1,7 @@
+// end the request
+function end(res, text) {
+  console.log(text);
+  res.status(200).send(text);
+}
+
+export default end;
